requirements.txt
snowflake-connector-python
setuptools